+++ lib/Runtime/ByteCode/AsmJsByteCodeWriter.cpp
@@ -67,7 +67,7 @@ namespace Js
         return offset;
     }
 
-    void AsmJsByteCodeWriter::InitData(ArenaAllocator* alloc, long initCodeBufferSize)
+    void AsmJsByteCodeWriter::InitData(ArenaAllocator* alloc, int32 initCodeBufferSize)
     {
         ByteCodeWriter::InitData(alloc, initCodeBufferSize);
 #ifdef BYTECODE_BRANCH_ISLAND
