+++ lib/Runtime/Library/DateImplementation.cpp
@@ -34,7 +34,7 @@ namespace Js {
         const char16 *psz;      // string
         short cch;              // length of string
         short szst;             // type of entry
-        long lwVal;             // value
+        int32 lwVal;             // value
     };
 
     BEGIN_ENUM_BYTE(ParseStringTokenType)
@@ -109,7 +109,7 @@ namespace Js {
 #undef szst
 #undef Szs
     };
-    const long kcszs = sizeof(g_rgszs) / sizeof(SZS);
+    const int32 kcszs = sizeof(g_rgszs) / sizeof(SZS);
 
     // Moved DaylightTimeHelper to common.lib to share with hybrid debugging, but this function depends on runtime.
     bool DaylightTimeHelper::ForceOldDateAPIFlag()
@@ -415,7 +415,7 @@ namespace Js {
             const charcount_t cchWritten = NumberUtilities::UInt16ToString(value, buffer, charCapacity, 2);
             Assert(cchWritten != 0);
         };
-        const auto ConvertLongToString = [](const long value, char16 *const buffer, const CharCount charCapacity)
+        const auto ConvertLongToString = [](const int32 value, char16 *const buffer, const CharCount charCapacity)
         {
             const errno_t err = _ltow_s(value, buffer, charCapacity, 10);
             Assert(err == 0);
@@ -1072,18 +1072,18 @@ Error:
         char16 ch;
         char16 *pszSrc = nullptr;
 
-        const long lwNil = 0x80000000;
-        long cch;
-        long depth;
-        long lwT;
-        long lwYear = lwNil;
-        long lwMonth = lwNil;
-        long lwDate = lwNil;
-        long lwTime = lwNil;
-        long lwZone = lwNil;
-        long lwOffset = lwNil;
+        const int32 lwNil = 0x80000000;
+        int32 cch;
+        int32 depth;
+        int32 lwT;
+        int32 lwYear = lwNil;
+        int32 lwMonth = lwNil;
+        int32 lwDate = lwNil;
+        int32 lwTime = lwNil;
+        int32 lwZone = lwNil;
+        int32 lwOffset = lwNil;
 
-        long ss = ssNil;
+        int32 ss = ssNil;
         const SZS *pszs;
 
         bool fUtc;
@@ -1165,7 +1165,7 @@ Error:
                 for ( ; !FBig(*pch) && (isalpha(*pch) || '.' == *pch); pch++)
                     ;
 
-                cch = (long)(pch - pchBase);
+                cch = (int32)(pch - pchBase);
 
                 if ('.' == pchBase[cch - 1])
                 {
@@ -1199,11 +1199,11 @@ Error:
                         {
                             goto LError;
                         }
-                        lwZone = -(long)(ch - 'a' + (ch < 'j')) * 60;
+                        lwZone = -(int32)(ch - 'a' + (ch < 'j')) * 60;
                     }
                     else if (ch <= 'y')
                     {
-                        lwZone = (long)(ch - 'm') * 60;
+                        lwZone = (int32)(ch - 'm') * 60;
                     }
                     else if (ch == 'z')
                     {
